Rename module WaveGenerator to WaveManager
