--- WaveGenerator/src/main/java/module-info.java
@@ -1,2 +1,0 @@
-module WaveGenerator {
-}
