+++ WaveManager/src/main/java/com/github/sef24sp4/wavemanager/IWaveManager.java
@@ -1,4 +1,4 @@
-package com.github.sef24sp4.wavegenerator;
+package com.github.sef24sp4.wavemanager;
 
 /**
  * This interface is not done yet.
