+++ WaveManager/src/main/java/module-info.java
@@ -1,0 +1,2 @@
+module WaveManager {
+}
